requirements.txt
fastapi
pydantic
motor
python-dotenv
openai
